refactor: Signature checking using function stored in attributes
